Get serialED sim and indexing benchmark scripts working again with modern DIALS

+++ dgw/indigo_benchmark.py
@@ -5,7 +5,8 @@ Script to compare various methods of indexing ED stills
 
 import os
 import glob
-from libtbx import easy_run
+import subprocess
+import shutil
 from dials.array_family import flex
 from libtbx.table_utils import simple_table
 from dxtbx.model.experiment_list import ExperimentListFactory
@@ -19,24 +20,39 @@ class Script(object):
 
         self.cmds = [
             (
-                "dials.index stills.indexer=sweeps imported.expt "
-                "strong.refl unit_cell=78.840,78.840,38.290,90.000,90.000,90.000 "
-                "space_group=P43212 indexing.method=real_space_grid_search "
-                "n_macro_cycles=2 detector.fix_list=Dist"
+                shutil.which("dials.index"),
+                "imported.expt",
+                "strong.refl",
+                "stills.indexer=sequences",
+                "unit_cell=78.840,78.840,38.290,90.000,90.000,90.000",
+                "space_group=P43212",
+                "indexing.method=real_space_grid_search",
+                "n_macro_cycles=2",
+                "detector.fix=distance",
             ),  # RSGS
             (
-                "dials.index imported.expt strong.refl "
-                "unit_cell=78.840,78.840,38.290,90.000,90.000,90.000 "
-                "space_group=P43212 indexing.method=low_res_spot_match "
-                "stills.indexer=sweeps n_macro_cycles=2 detector.fix_list=Dist "
-            ),  # indigo
+                shutil.which("dials.index"),
+                "imported.expt",
+                "strong.refl",
+                "stills.indexer=sequences",
+                "unit_cell=78.840,78.840,38.290,90.000,90.000,90.000",
+                "space_group=P43212",
+                "indexing.method=low_res_spot_match",
+                "n_macro_cycles=2",
+                "detector.fix=distance",
+            ),  # lrsm
             (
-                "dials.index imported.expt strong.refl "
-                "unit_cell=78.840,78.840,38.290,90.000,90.000,90.000 "
-                "space_group=P43212 indexing.method=low_res_spot_match "
-                "stills.indexer=sweeps n_macro_cycles=2 detector.fix_list=Dist "
-                "bootstrap_crystal=True "
-            ),  # indigo + bootstrap crystal
+                shutil.which("dials.index"),
+                "imported.expt",
+                "strong.refl",
+                "stills.indexer=sequences",
+                "unit_cell=78.840,78.840,38.290,90.000,90.000,90.000",
+                "space_group=P43212",
+                "indexing.method=low_res_spot_match",
+                "n_macro_cycles=2",
+                "detector.fix=distance",
+                "bootstrap_crystal=True",
+            ),  # lrsm + bootstrap crystal
         ]
 
         self.images = sorted(glob.glob("noiseimage_*.img"))
@@ -89,37 +105,27 @@ class Script(object):
         return abs(angle)
 
     def process(self, image):
-        print("processing {0}".format(image))
-        cmd = "dials.import {0}".format(image)
-        result = easy_run.fully_buffered(command=cmd)
-        cmd = "dials.find_spots imported.expt min_spot_size=3"
-        result = easy_run.fully_buffered(command=cmd)
-        strong = flex.reflection_table.from_pickle("strong.refl")
+        print(f"processing {image}")
+        cmd = (shutil.which("dials.import"), image)
+        result = subprocess.run(cmd)
+        cmd = (shutil.which("dials.find_spots"), "imported.expt", "min_spot_size=3")
+        result = subprocess.run(cmd)
+        strong = flex.reflection_table.from_file("strong.refl")
         d = {"nspots": len(strong)}
 
         nindexed = []
         offset_deg = []
         for cmd in self.cmds:
-            success = True
-            try:
-                result = easy_run.fully_buffered(command=cmd).raise_if_errors()
-                assert result.return_code == 0
-            except RuntimeError:
-                success = False
-
-            if success:
+            result = subprocess.run(cmd)
+            if os.path.isfile("indexed.refl"):
                 idx = flex.reflection_table.from_file("indexed.refl")
                 nindexed.append(idx.get_flags(idx.flags.indexed).count(True))
                 offset_deg.append(self.compare_orientation_matrices(image))
+                os.remove("indexed.refl")
+                os.remove("indexed.expt")
             else:
                 nindexed.append(0)
                 offset_deg.append(None)
-
-            try:
-                os.remove("indexed.refl")
-                os.remove("indexed.expt")
-            except OSError:
-                pass
 
         d["nindexed"] = nindexed
         d["offset_deg"] = offset_deg
